Module Wodan

The Wodan module allows access to Wodan filesystems

The entry point is Wodan.Make, which returns a Wodan.S module supporting filesystem operations

exception BadMagic

Raised when trying to open a superblock and it doesn't have the magic tag for Wodan filesystems

exception BadVersion

Raised when trying to open a superblock and it doesn't have a supported version for Wodan filesystems

exception BadFlags

Raised when trying to open a superblock and it sets must-support flags that this version doesn't support

exception BadCRC of int64

Raised when a block doesn't have the expected CRC

The exception carries the logical offset of the block.

exception BadParams

Raised when SUPERBLOCK_PARAMS passed to the Make functor don't match actual settings saved in the superblock

exception ReadError

Raised on IO failures reading a block

exception WriteError

Raised on IO failures writing a block

exception OutOfSpace

Raised when the filesystem doesn't have enough space to perform the operation

exception NeedsFlush

Raised when the filesystem doesn't have enough space to perform an operation that would be possible if pending operations were flushed first

exception BadKey of string

Raised when converting to a key is impossible, eg because the input doesn't have the expected length

The exception carries the original input.

exception ValueTooLarge of string

Raised when converting to a value is impossible, because the input would not fit on a single block

Note that Wodan expects the user to chunk values if arbitrary lengths are to be supported. See also Wodan_irmin's relationship to irmin-chunk.

The exception carries the original input

exception BadNodeType of int

Raised when a block doesn't have a known and expected node type

The exception carries the type tag that couldn't be handled.

module type EXTBLOCK = sig ... end

The standard Mirage block signature, with Wodan-specific extensions

module BlockCompat : functor (B : Mirage_block.S) -> EXTBLOCK with type BlockCompat.t = B.t

Extend a basic Mirage block backend to provide Wodan-specific extensions

val sizeof_superblock : int

A constant that represents the size of Wodan superblocks

Will be a power of two and a multiple of a standard sector size

type relax = {

magic_crc : bool;
CRC errors ignored on any read where the magic CRC is used

magic_crc_write : bool;
Write non-superblock blocks with the magic CRC

}

Flags representing integrity features that will be relaxed

Passed at mount time through mount_options

type mount_options = {

fast_scan : bool;
If enabled, instead of checking the entire filesystem when opening, leaf nodes won't be scanned. They will be scanned on open instead.

cache_size : int;
How many blocks to keep in cache

relax : relax;
Integrity invariants to relax

}

All parameters that can't be read from the superblock

module OptionalSuperblockFlags : sig ... end

Superblock flags that are supported but not required by this Wodan version

module type SUPERBLOCK_PARAMS = sig ... end

All parameters that can be read from the superblock

module Testing : sig ... end

Operations used when testing and fuzzing

module StandardSuperblockParams : SUPERBLOCK_PARAMS

Defaults for SUPERBLOCK_PARAMS

val standard_mount_options : mount_options

Defaults for mount_options

val read_superblock_params : (module Mirage_block.S with type t = 'a) -> 'a -> relax -> (module SUPERBLOCK_PARAMS) Lwt.t

Read static filesystem parameters

These are set at creation time and recorded in the superblock. See open_for_reading if all you need is to mount the filesystem.

type format_params = {

logical_size : int64;
The number of blocks, including the superblock, that are part of the filesystem

preroots_interval : int64;
The (maximum) interval between pre-roots. Pre-roots are used to ensure that bisection will not be slow on freshly-formatted devices.

}

Parameters passed when creating a filesystem

val default_preroots_interval : int64

type deviceOpenMode =

| OpenExistingDevice
Open an existing device, read logical size from superblock

| FormatEmptyDevice of format_params
Format a device, which must contain only zeroes, and use the given format_params

Modes for opening a device, see S.prepare_io

module type S = sig ... end

Filesystem operations

module Make : functor (B : EXTBLOCK) -> functor (P : SUPERBLOCK_PARAMS) -> S with type disk = B.t

Build a Wodan.S module given a backing device and parameters

type open_ret =

| OPEN_RET : (module S with type root = 'a) * 'a * int64 -> open_ret

This is a type that packages together a Wodan.S with an opened root and the generation number read when mounting

val open_for_reading : (module EXTBLOCK with type EXTBLOCK.t = 'a) -> 'a -> mount_options -> open_ret Lwt.t

Open an existing Wodan filesystem, getting static parameters from the superblock