Module Wodan.OptionalSuperblockFlags

Superblock flags that are supported but not required by this Wodan version

type t

A set of flags

val empty : t

The empty set of flags

val tombstones_enabled : t

The flag for tombstone support

If set, the empty value will be considered a tombstone, meaning that functions that query values (mem, lookup, iter, search_range…) will treat it as if there was no value.

This enables optimisations on the write path, allowing tombstones to fully disappear from storage eventually

val intersect : t -> t -> t

Take two sets of flags, return the common subset