Module type Wodan.S

Filesystem operations

This module is specialized for a given set of superblock parameters, with key and value types also being specialized to match.

type key

An opaque type for fixed-size keys

Conversion from/to strings is free

type value

An opaque type for bounded-size values

Conversion from/to strings is free

type disk

A backing device

type root

A filesystem root

module Key : sig ... end

Operations over keys

module P : SUPERBLOCK_PARAMS

The parameter module that was used to create this module

val key_of_cstruct : Cstruct.t -> key

val key_of_string : string -> key

val key_of_string_padded : string -> key

val cstruct_of_key : key -> Cstruct.t

val string_of_key : key -> string

val value_of_cstruct : Cstruct.t -> value

val value_of_string : string -> value

val value_equal : value -> value -> bool

val cstruct_of_value : value -> Cstruct.t

val string_of_value : value -> string

val next_key : key -> key

The next highest key

Raises Invalid_argument if already at the highest possible key

val is_tombstone : value -> bool

Whether a value is a tombstone

val insert : root -> key -> value -> unit Lwt.t

Store data in the filesystem

Any previously stored value will be silently overwritten

val lookup : root -> key -> value option Lwt.t

Read data from the filesystem

val mem : root -> key -> bool Lwt.t

Check whether a key has been set within the filesystem

val flush : root -> int64 Lwt.t

Send changes to disk

val fstrim : root -> int64 Lwt.t

Discard all blocks which the filesystem doesn't explicitly use

val live_trim : root -> int64 Lwt.t

Discard blocks that have been unused since mounting or since the last live_trim call

val log_statistics : root -> unit

Send statistics about operations to the log

val search_range : root -> key -> key -> (key -> value -> unit) -> unit Lwt.t

Call back a function for all elements in the range from start inclusive to end_ exclusive

Results are in no particular order.

val iter : root -> (key -> value -> unit) -> unit Lwt.t

Call back a function for all elements in the filesystem

val prepare_io : deviceOpenMode -> disk -> mount_options -> (root * int64) Lwt.t

Open a filesystem

Returns a root and its generation number. When integrating Wodan as part of a distributed system, you may want to check here that the generation number has grown since the last flush