Module Wodan__Bitv64

Bitv64 is a wrapper around Bitv, for easier use with Location.t indexes and int64 sizes.

type t

The type for Location.t indexed Bitvs

val create : int64 -> bool -> t

create n b creates a new bit vector of length n, initialized with b. n must be smaller than max_int, otherwise raises Invalid_argument.

val set : t -> Wodan__.Location.t -> bool -> unit

Bitv.set v n b sets the nth bit of v to the value b.

val get : t -> Wodan__.Location.t -> bool

Bitv.get v n returns the nth bit of v.

val length : t -> int64

length returns the length of the given vector.

val iter : (bool -> unit) -> t -> unit

iter f v applies f to every element in v.

val iteri : (Wodan__.Location.t -> bool -> unit) -> t -> unit

iteri is like iter, but applies f to the index of the element, as first argument, and to the element itself, as second argument.